Store current DS block, final block, and states into metadata
