+++ src/libData/BlockData/BlockHeader/DSBlockHeader.h
@@ -25,7 +25,7 @@
 #include "common/Serializable.h"
 #include "libCrypto/Schnorr.h"
 #include "libData/AccountData/Transaction.h"
-#include "libUtils/UpgradeManager.h"
+#include "libUtils/SWInfo.h"
 
 /// Stores information on the header part of the DS block.
 class DSBlockHeader : public BlockHeaderBase
