+++ src/libUtils/UpgradeManager.cpp
@@ -23,114 +23,6 @@ using namespace std;
 #define RELEASE_URL                                                            \
     "https://api.github.com/repos/Zilliqa/Zilliqa/releases/latest"
 
-SWInfo::SWInfo()
-    : m_major(0)
-    , m_minor(0)
-    , m_fix(0)
-    , m_upgradeDS(0)
-    , m_commit(0)
-{
-}
-
-SWInfo::SWInfo(const uint32_t& major, const uint32_t& minor,
-               const uint32_t& fix, const uint64_t& upgradeDS,
-               const uint32_t& commit)
-    : m_major(major)
-    , m_minor(minor)
-    , m_fix(fix)
-    , m_upgradeDS(upgradeDS)
-    , m_commit(commit)
-{
-}
-
-SWInfo::SWInfo(const SWInfo& src)
-    : m_major(src.m_major)
-    , m_minor(src.m_minor)
-    , m_fix(src.m_fix)
-    , m_upgradeDS(src.m_upgradeDS)
-    , m_commit(src.m_commit)
-{
-}
-
-SWInfo::~SWInfo(){};
-
-/// Implements the Serialize function inherited from Serializable.
-unsigned int SWInfo::Serialize(std::vector<unsigned char>& dst,
-                               unsigned int offset) const
-{
-    LOG_MARKER();
-
-    unsigned int size_remaining = dst.size() - offset;
-
-    if (size_remaining < SIZE)
-    {
-        dst.resize(SIZE + offset);
-    }
-
-    unsigned int curOffset = offset;
-
-    SetNumber<uint32_t>(dst, curOffset, m_major, sizeof(uint32_t));
-    curOffset += sizeof(uint32_t);
-    SetNumber<uint32_t>(dst, curOffset, m_minor, sizeof(uint32_t));
-    curOffset += sizeof(uint32_t);
-    SetNumber<uint32_t>(dst, curOffset, m_fix, sizeof(uint32_t));
-    curOffset += sizeof(uint32_t);
-    SetNumber<uint64_t>(dst, curOffset, m_upgradeDS, sizeof(uint64_t));
-    curOffset += sizeof(uint64_t);
-    SetNumber<uint32_t>(dst, curOffset, m_commit, sizeof(uint32_t));
-    curOffset += sizeof(uint32_t);
-
-    return SIZE;
-}
-
-/// Implements the Deserialize function inherited from Serializable.
-int SWInfo::Deserialize(const std::vector<unsigned char>& src,
-                        unsigned int offset)
-{
-    LOG_MARKER();
-
-    unsigned int curOffset = offset;
-
-    try
-    {
-        m_major = GetNumber<uint32_t>(src, curOffset, sizeof(uint32_t));
-        curOffset += sizeof(uint32_t);
-        m_minor = GetNumber<uint32_t>(src, curOffset, sizeof(uint32_t));
-        curOffset += sizeof(uint32_t);
-        m_fix = GetNumber<uint32_t>(src, curOffset, sizeof(uint32_t));
-        curOffset += sizeof(uint32_t);
-        m_upgradeDS = GetNumber<uint64_t>(src, curOffset, sizeof(uint64_t));
-        curOffset += sizeof(uint64_t);
-        m_commit = GetNumber<uint32_t>(src, curOffset, sizeof(uint32_t));
-        curOffset += sizeof(uint32_t);
-    }
-    catch (const std::exception& e)
-    {
-        LOG_GENERAL(WARNING,
-                    "Error with SWInfo::Deserialize." << ' ' << e.what());
-        return -1;
-    }
-
-    return 0;
-}
-
-/// Less-than comparison operator.
-bool SWInfo::operator<(const SWInfo& r) const
-{
-    return tie(m_major, m_minor, m_fix, m_upgradeDS, m_commit)
-        < tie(r.m_major, r.m_minor, r.m_fix, r.m_upgradeDS, r.m_commit);
-}
-
-/// Greater-than comparison operator.
-bool SWInfo::operator>(const SWInfo& r) const { return r < *this; }
-
-/// Equality operator.
-bool SWInfo::operator==(const SWInfo& r) const
-{
-    return tie(m_major, m_minor, m_fix, m_upgradeDS, m_commit)
-        == tie(r.m_major, r.m_minor, r.m_fix, r.m_upgradeDS, r.m_commit);
-}
-
 UpgradeManager::UpgradeManager() {}
 
 UpgradeManager::~UpgradeManager() {}
@@ -293,12 +185,28 @@ bool UpgradeManager::DownloadSW()
     return true;
 }
 
-bool UpgradeManager::ReplaceNode()
+bool UpgradeManager::ReplaceNode(Mediator& mediator)
 {
     LOG_MARKER();
 
-    /// Store all the useful states into metadata, create a new node with loading the metadata, and kill current node
-    /// TBD
+    /// Store states
+    AccountStore::GetInstance().UpdateStateTrieAll();
+    AccountStore::GetInstance().MoveUpdatesToDisk();
 
-    return true;
+    /// Store DS block
+    vector<unsigned char> serializedDSBlock;
+    mediator.m_dsBlockChain.GetLastBlock().Serialize(serializedDSBlock, 0);
+    BlockStorage::GetBlockStorage().PutDSBlock(
+        mediator.m_dsBlockChain.GetLastBlock().GetHeader().GetBlockNum(),
+        serializedDSBlock);
+
+    /// Store final block
+    vector<unsigned char> serializedTxBlock;
+    mediator.m_txBlockChain.GetLastBlock().Serialize(serializedTxBlock, 0);
+    BlockStorage::GetBlockStorage().PutTxBlock(
+        mediator.m_txBlockChain.GetLastBlock().GetHeader().GetBlockNum(),
+        serializedTxBlock);
+
+    /// Call a script to kill current node, deploy downloaded software, and create a new node
+    return system("./replace.sh") >= 0;
 }
